requirements.txt
pandas
numpy
scikit-learn>=1.2
xgboost
lightgbm
catboost
matplotlib
mlflow
optuna
shap
joblib
pydantic
fastapi
uvicorn
hydra-core
pyyaml
click
streamlit
pytest
black
isort
flake8
pre-commit
python-dotenv
